+++ src/base/MidiDevice.cpp
@@ -791,7 +791,7 @@ MidiDevice::getIPBControlParameters() const
 ControlParameter *
 MidiDevice::getControlParameter(int index)
 {
-    if (index >= 0 && ((unsigned int)index) < m_controlList.size())
+    if (index >= 0 && ((unsigned int)index) < (unsigned int)m_controlList.size())
         return &m_controlList[index];
 
     return 0;
