+++ src/commands/edit/RetrogradeCommand.cpp
@@ -69,7 +69,7 @@ RetrogradeCommand::modifySegment()
 
     Segment &segment(m_selection->getSegment());
 
-    for (unsigned int j = 0; j < toErase.size(); ++j) {
+    for (size_t j = 0; j < toErase.size(); ++j) {
         Segment::iterator jtr(segment.findSingle(toErase[j]));
         if (jtr != segment.end()) {
             RG_DEBUG << "found event " << j << endl;
@@ -79,7 +79,7 @@ RetrogradeCommand::modifySegment()
         }
     }
 
-    for (unsigned int j = 0; j < toInsert.size(); ++j) {
+    for (size_t j = 0; j < toInsert.size(); ++j) {
 
         Segment::iterator jtr = segment.end();
