+++ src/commands/segment/SegmentSyncCommand.cpp
@@ -47,7 +47,7 @@ SegmentSyncCommand::SegmentSyncCommand(SegmentSelection selection, int newTransp
 SegmentSyncCommand::SegmentSyncCommand(std::vector<Segment *> segments, int newTranspose, int lowRange, int highRange, const Clef& clef) :
         MacroCommand(tr("Sync segment parameters"))
 {
-	for (int i = 0; i < segments.size(); i++) {
+	for (size_t i = 0; i < segments.size(); i++) {
         processSegment(*(segments[i]), newTranspose, lowRange, highRange, clef);
     }
 }
